requirements.txt
streamlit
transformers
torch
pillow
flask
transformers 
torch 
torchvision 
torchaudio 
pillow 
